third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-font-expected.txt

This test exercises the 'font' shorthand property in CSS computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedFont('10px sans-serif') is 'normal normal 400 normal 10px / normal sans-serif'
PASS computedFont('10px sans-serif') is 'normal normal 400 normal 10px / normal sans-serif'
PASS computedFont('10px SANS-SERIF') is 'normal normal 400 normal 10px / normal sans-serif'
PASS computedFont('12px sans-serif') is 'normal normal 400 normal 12px / normal sans-serif'
PASS computedFont('12px  sans-serif') is 'normal normal 400 normal 12px / normal sans-serif'
PASS computedFont('10px sans-serif, sans-serif') is 'normal normal 400 normal 10px / normal sans-serif, sans-serif'
PASS computedFont('10px sans-serif, serif') is 'normal normal 400 normal 10px / normal sans-serif, serif'
PASS computedFont('12px ahem') is 'normal normal 400 normal 12px / normal ahem'
PASS computedFont('12px unlikely-font-name') is 'normal normal 400 normal 12px / normal unlikely-font-name'
PASS computedFont('100 10px sans-serif') is 'normal normal 100 normal 10px / normal sans-serif'
PASS computedFont('200 10px sans-serif') is 'normal normal 200 normal 10px / normal sans-serif'
PASS computedFont('300 10px sans-serif') is 'normal normal 300 normal 10px / normal sans-serif'
PASS computedFont('400 10px sans-serif') is 'normal normal 400 normal 10px / normal sans-serif'
PASS computedFont('normal 10px sans-serif') is 'normal normal 400 normal 10px / normal sans-serif'
PASS computedFont('500 10px sans-serif') is 'normal normal 500 normal 10px / normal sans-serif'
PASS computedFont('600 10px sans-serif') is 'normal normal 600 normal 10px / normal sans-serif'
FAIL computedFont('700 10px sans-serif') should be normal normal bold normal 10px / normal sans-serif. Was normal normal 700 normal 10px / normal sans-serif.
FAIL computedFont('bold 10px sans-serif') should be normal normal bold normal 10px / normal sans-serif. Was normal normal 700 normal 10px / normal sans-serif.
PASS computedFont('800 10px sans-serif') is 'normal normal 800 normal 10px / normal sans-serif'
PASS computedFont('900 10px sans-serif') is 'normal normal 900 normal 10px / normal sans-serif'
PASS computedFont('italic 10px sans-serif') is 'italic normal 400 normal 10px / normal sans-serif'
PASS computedFont('small-caps 10px sans-serif') is 'normal small-caps 400 normal 10px / normal sans-serif'
PASS computedFont('italic small-caps 10px sans-serif') is 'italic small-caps 400 normal 10px / normal sans-serif'
PASS computedFont('italic small-caps bold 10px sans-serif') is 'italic small-caps 700 normal 10px / normal sans-serif'
PASS computedFont('10px/100% sans-serif') is 'normal normal 400 normal 10px / 10px sans-serif'
PASS computedFont('10px/100px sans-serif') is 'normal normal 400 normal 10px / 100px sans-serif'
PASS computedFont('10px/normal sans-serif') is 'normal normal 400 normal 10px / normal sans-serif'
PASS computedFont('10px/normal sans-serif') is 'normal normal 400 normal 10px / normal sans-serif'
PASS successfullyParsed is true

TEST COMPLETE
